+++ app/components/files/file-inline-image.js
@@ -204,7 +204,8 @@ export default class FileInlineImage extends SafeComponent {
             padding: this.outerPadding,
             borderColor: vars.lightGrayBg,
             borderWidth: 1,
-            marginVertical: 4
+            marginVertical: 4,
+            borderRadius: 2
         };
 
         const header = {
@@ -217,12 +218,15 @@ export default class FileInlineImage extends SafeComponent {
             flexGrow: 1,
             flexShrink: 1,
             fontWeight: 'bold',
-            color: vars.txtMedium
+            color: vars.txtMedium,
+            textAlignVertical: 'top',
+            lineHeight: 25
         };
 
         const titleText = {
             color: vars.bg,
-            marginVertical: 2
+            marginVertical: 2,
+            ellipsizeMode: 'tail'
         };
 
         const descText = {
@@ -242,7 +246,7 @@ export default class FileInlineImage extends SafeComponent {
                         {!!description && <Text style={descText}>{description}</Text>}
                     </View>
                     <View style={header}>
-                        {!!name && <Text style={text}>{name}</Text>}
+                        {!!name && <Text numberOfLines={1} ellipsizeMode="tail" style={text}>{name}</Text>}
                         {isLocal && <View style={{ flexDirection: 'row' }}>
                             {!downloading && icons.darkNoPadding(this.opened ? 'arrow-drop-up' : 'arrow-drop-down', () => { this.opened = !this.opened; })}
                             {!downloading && icons.darkNoPadding('more-vert', () => this.props.onAction(this.props.image))}