+++ app/components/login/login-state.js
@@ -75,7 +75,6 @@ class LoginState extends RoutedState {
     }
 
     @action _login(user) {
-        console.log(`login-state.js: logging in`);
         User.current = user;
         return user.login()
             .then(() => console.log('login-state.js: logged in'))
@@ -86,12 +85,17 @@ class LoginState extends RoutedState {
                 await this.enableAutomaticLogin(user);
             })
             .catch(e => {
+                this.isInProgress = false;
                 console.error(e);
                 User.current = null;
                 this.passphraseValidationMessage = tx('error_wrongAK');
                 return Promise.reject(new Error(this.error));
-            })
-            .then(() => mainState.activateAndTransition(user))
+            });
+    }
+
+    transition() {
+        const user = User.current;
+        return new Promise(() => mainState.activateAndTransition(user))
             .then(() => this.clean())
             .then(async () => {
                 if (this._resetTouchId) {