+++ app/components/messaging/chat-item.js
@@ -28,6 +28,7 @@ export default class ChatItem extends SafeComponent {
         const error = !!i.signatureError;
         const systemMessageText =
             i.systemData && systemMessages.getSystemMessageText(i) || null;
+        const videoCallLink = i.systemData && i.systemData.link || null;
         const files = i.files && i.files.map(id => fileState.store.getById(id)).filter(f => f) || [];
         const images = files.filter(f => f.isImage) || [];
         const normalFiles = files.filter(f => !f.isImage) || [];
@@ -56,6 +57,7 @@ export default class ChatItem extends SafeComponent {
                     isChat
                     fullnameIsBold
                     systemMessage={systemMessageText}
+                    videoCallMessage={videoCallLink}
                     key={key}
                     error={error}
                     onPress={onPress}