+++ app/lib/mobile-config.js
@@ -16,6 +16,7 @@ export default (c, icebear) => {
         tfaDetail: 'https://peerio.zendesk.com/hc/en-us/articles/203665635-What-is-two-factor-authentication-',
         msgSignature: 'https://peerio.zendesk.com/hc/en-us/articles/204394135',
         upgrade: 'route:modal:accountUpgradeSwiper',
+        createRoom: 'route:modal:createChannel',
         signup: 'route:app:signupStep1',
         settings: 'route:main:settings',
         proWelcome: 'https://peerio.zendesk.com/hc/en-us/articles/208395556',
@@ -45,7 +46,8 @@ export default (c, icebear) => {
     cfg.upload.uploadBufferSize = 1024 * 1024;
 
     cfg.isMobile = true;
-    cfg.socketServerUrl = process.env.PEERIO_SOCKET_SERVER || 'wss://icebear.peerio.com';
+    // socket server is always taken from env
+    cfg.socketServerUrl = process.env.PEERIO_SOCKET_SERVER || 'wss://changeme.peerio.com';
     cfg.FileStream = rnFileStream(icebear.FileStreamAbstract);
     cfg.StorageEngine = KeyValueStorage;